refactor(build-logic): extract dependency resolution strategy into shared helper
