+++ build-logic/src/main/kotlin/AndroidConfiguration.kt
@@ -12,7 +12,6 @@ internal fun LibraryExtension.baseConfiguration(project: Project) {
     configureSdk(project)
     configureProguardRules(project)
     configureTestOptions(project)
-    project.configureDependencyResolutionStrategy()
 }
 
 private fun BaseExtension.configureSdk(project: Project) {
@@ -21,25 +20,6 @@ private fun BaseExtension.configureSdk(project: Project) {
     defaultConfig {
         minSdk = project.versionCatalog.findVersion("android-sdk-min").get().requiredVersion.toInt()
         targetSdk = project.versionCatalog.findVersion("android-sdk-compile").get().requiredVersion.toInt()
-    }
-}
-
-private fun Project.configureDependencyResolutionStrategy() {
-    configurations.configureEach {
-        resolutionStrategy {
-            // HACK: force androidx-annotation version for several modules
-            //       known modules requiring the forced version: androidx-constraintlayout, androidx-core
-            force(versionCatalog.findLibrary("androidx-annotation").get())
-
-            // use the new condensed version of hamcrest
-            dependencySubstitution {
-                val hamcrestVersion = versionCatalog.findVersion("hamcrest").get().requiredVersion
-                substitute(module("org.hamcrest:hamcrest-core"))
-                    .using(module("org.hamcrest:hamcrest:$hamcrestVersion"))
-                substitute(module("org.hamcrest:hamcrest-library"))
-                    .using(module("org.hamcrest:hamcrest:$hamcrestVersion"))
-            }
-        }
     }
 }
