+++ build-logic/src/main/kotlin/gto-support.android-conventions.gradle.kts
@@ -20,4 +20,5 @@ dependencies {
     testImplementation(versionCatalog.findBundle("android-test-framework").get())
 }
 
+baseDependencyResolutionStrategy()
 configureTestSharding()
