+++ build-logic/src/main/kotlin/gto-support.java-conventions.gradle.kts
@@ -10,3 +10,5 @@ plugins {
 kotlin {
     configureJvmToolchain(project)
 }
+
+baseDependencyResolutionStrategy()
