+++ build-logic/src/main/kotlin/gto-support.multiplatform-conventions.gradle.kts
@@ -16,3 +16,5 @@ kotlin {
         }
     }
 }
+
+baseDependencyResolutionStrategy()
